refactor: wish/item 타입 전역으로 추출
