+++ apps/web/src/app/item/[id]/edit/_apis/getWish.ts
@@ -4,14 +4,18 @@ import { clientApi } from '@/apis/client';
 import { serverApi } from '@/apis/server';
 import type { ApiResponseT } from '@/types/api';
 
-import type { WishT } from '../_types/wish';
+import type { GetWishResponseT } from '../_types/wish';
 
 export const getWish = async (wishId: number) => {
   if (environmentManager.isServer()) {
-    const { data } = await serverApi.get<ApiResponseT<WishT>>(`/api/v1/wishlists/${wishId}`);
+    const { data } = await serverApi.get<ApiResponseT<GetWishResponseT>>(
+      `/api/v1/wishlists/${wishId}`
+    );
     return data.data;
   }
 
-  const { data } = await clientApi.get<ApiResponseT<WishT>>(`/api/v1/wishlists/${wishId}`);
+  const { data } = await clientApi.get<ApiResponseT<GetWishResponseT>>(
+    `/api/v1/wishlists/${wishId}`
+  );
   return data.data;
 };
