+++ apps/web/src/types/wish.ts
@@ -1,0 +1,4 @@
+export type WishT = {
+  id: number;
+  createdAt: string;
+};
